+++ app/src/main/java/com/madonasyombua/sportsdb/data/remote/response/TeamsResponse.kt
@@ -1,0 +1,9 @@
+package com.madonasyombua.sportsdb.data.remote.response
+
+import com.madonasyombua.sportsdb.data.remote.model.TeamByLeague
+import com.squareup.moshi.JsonClass
+
+@JsonClass(generateAdapter=true)
+data class TeamsResponse(
+    val teams:List<TeamByLeague>
+)
